allowing rows to be saved after each run

+++ DrivewayDentDeletion/continuous-load.sh
@@ -15,6 +15,7 @@
 #   -t : <retry_interval>, (integer), time in seconds between each load of data
 #   -c : <should_cleanup_table> (true/false), whether to delete all rows from the test table
 #   -i : <condensed_info> (true/false), whether to show the full post response or a condensed version
+#   -s : <save_row_after_run> (true/false), whether to save each row in the database after a run or delete it
 #
 # USAGE:
 #   CAUTION - running without <should_cleanup_table> enabled can result in data leftover in the postgres table
@@ -26,15 +27,16 @@
 #     ./continuous-load.sh -t 2 -c
 
 function usage {
-    echo "Usage: $0 -a <api_base_url> -t <retry_interval> -c <should_cleanup_table> -i <condensed_info>"
+    echo "Usage: $0 -a <api_base_url> -t <retry_interval> -c -i -s"
 }
 
 should_cleanup_table=false
 condensed_info=false
+save_row_after_run=false
 api_base_url=$(echo "http://$(oc get routes -n ace | grep ace-ddd-api-dev-http-ace | awk '{print $2}')/drivewayrepair")
 retry_interval=5
 
-while getopts "a:t:ci" opt; do
+while getopts "a:t:cis" opt; do
   case ${opt} in
     a ) api_base_url="$OPTARG"
       ;;
@@ -43,6 +45,8 @@ while getopts "a:t:ci" opt; do
     c ) should_cleanup_table=true
       ;;
     i ) condensed_info=true
+      ;;
+    s ) save_row_after_run=true
       ;;
     \? ) usage
       ;;
@@ -104,10 +108,12 @@ while true; do
     fi
 
     # - DELETE ---
-    echo -e "\nDeleting row from database..."
-    oc exec -n postgres -it $(oc get pod -n postgres -l name=postgresql -o jsonpath='{.items[].metadata.name}') \
-      -- psql -U admin -d sampledb -c \
-    "DELETE FROM quotes WHERE quotes.quoteid = ${quote_id};"
+    if [ "$save_row_after_run" = false ]; then
+      echo -e "\nDeleting row from database..."
+      oc exec -n postgres -it $(oc get pod -n postgres -l name=postgresql -o jsonpath='{.items[].metadata.name}') \
+        -- psql -U admin -d sampledb -c \
+      "DELETE FROM quotes WHERE quotes.quoteid = ${quote_id};"
+    fi
   else
     echo "FAILED - Error code: ${post_response_code}" # Failure catch during POST
   fi
